Add direct link to reply/comment
It's difficult to be able to copy and paste a link to a specific
comment or reply. It's only easy from the notifications page.

This patch introduces a link to a comment/reply by changing
the <span> to a <a> around the "When" text like "15 hours ago".

This is usually how other websites do deep links to comments too,
they make the time of post into a clickable link, so people should
be able to find it easyly.