Add community icon infront of name in teasers
I have a much easier time recognizing icons, even if they're small
compared to reading the name of the community. I thought for
example that beehive blocked me because I never saw any posts from
there.

Turns out I see all of them, I just didn't realize that those posts
were from beehive, because we don't show the domain part of a
community, so I assumed every time I saw c/technology it came
from lemme world. With the icon I can see imediatelly that it's
from beehive.

It does make the page a bit heavier because the browser needs to
download all the community icons, but in my opinion it is worth
it for the better UX.