Add visible lemmy_link to feed page
When I want to remotely subscribe to a piefed community then on
the community page there is the name in plain text for me to
copy and paste like: !community@example.com

But when I want to remotely subscribe to a feed then there is no
such url/name available on the feed page. It is nice when
this ~feedname@example.com is in the same place as the
community one.

I am aware that I can just put in the
URL https://example.com/f/feedname in and it will work, but I
think it is good for consistency.