After (un)subscribe to a community, add link to flash()
When you join a communyty through add_remote you stay on the
/community/add_remote page to be able to search for the next
one and also add it. But most of the time after I Join a
community I want to look at it. Ideally it would just redirect
me to it afterwards, but that is now how it works now.

You also get a flash() at the top to show you the success, so
I added a link to the name of the community so you can now click
on it and easily go to the community. I did the same for
leaving a community.