Add link to 'Comments in ...' headline
Now that we render comments from other communities under the post
it is convinient to have a link to the comments on the original
post.

This commit adds a link around the
'Comments in community@example.com' headline so that you easily
can go to the specific post page where the comments are from and
see the community rules, etc. in the sidebar.